feat: add market-table component
