+++ src/components/market/market-table/index.ts
@@ -1,0 +1,1 @@
+export { default } from './market-table';
